requirements.txt
appscript==1.2.0
click==8.1.3
docx2pdf==0.1.8
docxcompose==1.3.6
docxtpl==0.16.4
et-xmlfile==1.1.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.3
openpyxl==3.0.10
pandas==1.5.0
pathspec==0.10.1
platformdirs==2.5.2
PyQt6==6.4.0
PyQt6-Qt6==6.4.0
PyQt6-sip==13.4.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.4
six==1.16.0
tomli==2.0.1
tqdm==4.64.1
